requirements.txt
asgiref==3.2.7
attrs==19.3.0
boto3==1.12.40
botocore==1.15.40
CacheControl==0.12.6
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
Django==3.0.5
django-location-field==2.1.0
django-mapbox-location-field==1.4.0
django-storages==1.9.1
docutils==0.15.2
geographiclib==1.50
geopy==1.21.0
google-api-core==1.17.0
google-auth==1.14.1
google-cloud-core==1.3.0
google-cloud-storage==1.28.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
iso3166==1.0.1
jmespath==0.9.5
jsonschema==3.2.0
jsonseq==1.0.0
mapbox==0.16.1
mapbox-tilesets==1.0.1.dev0
mapboxcli==0.8.0
msgpack==1.0.0
polyline==1.4.0
protobuf==3.11.3
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-toolbelt==0.9.1
rsa==4.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.1.0
